NetEMD Network Earth Mover's Distance

Calculates the mean minimum Earth Mover's Distance (EMD) between two sets of discrete histograms after normalising each histogram to unit mass and variance. This is calculated as follows: 1. Normalise each histogram to have unit mass and unit variance 2. Find the minimum EMD between each pair of histograms 3. Take the average minimum EMD across all histogram pairs

net_emd(dhists1, dhists2, method = "optimise", return_details = FALSE,
  smoothing_window_width = 0)

Arguments

dhists1
A dhist discrete histogram object or a list of such objects

dhists2
A dhist discrete histogram object or a list of such objects

method
The method to use to find the minimum EMD across all potential offsets for each pair of histograms. Default is "optimise" to use R's built-in stats::optimise method to efficiently find the offset with the minimal EMD. However, this is not guaranteed to find the global minimum if multiple local minima EMDs exist. You can alternatively specify the "exhaustive" method, which will exhaustively evaluate the EMD between the histograms at all offsets that are candidates for the minimal EMD.

return_details
Logical indicating whether to return the individual minimal EMDs and associated offsets for all pairs of histograms

smoothing_window_width
Width of "top-hat" smoothing window to apply to "smear" point masses across a finite width in the real domain. Default is 0, which results in no smoothing. Care should be taken to select a smoothing_window_width that is appropriate for the discrete domain (e.g.for the integer domain a width of 1 is the natural choice)

Value

NetEMD measure for the two sets of discrete histograms (return_details = FALSE) or a list with the following named elements net_emd: the NetEMD for the set of histogram pairs, min_emds: the minimal EMD for each pair of histograms, min_offsets: the associated offsets giving the minimal EMD for each pair of histograms