Netdis comparisons between one graph and many other graphs.

Netdis comparisons between one graph and many other graphs.

netdis_one_to_many(
  graph_1 = NULL,
  graphs_compare = NULL,
  ref_graph = 0,
  max_graphlet_size = 4,
  neighbourhood_size = 2,
  min_ego_nodes = 3,
  min_ego_edges = 1,
  binning_fn = NULL,
  bin_counts_fn = NULL,
  exp_counts_fn = NULL,
  graphlet_counts_1 = NULL,
  graphlet_counts_compare = NULL,
  graphlet_counts_ref = NULL
)

Arguments

graph_1

Query graph - this graph will be compared with all graphs in graphs_compare. A simplified igraph graph object.

graphs_compare

Graphs graph_1 will be compared with. A named list of simplified igraph graph objects.

ref_graph

Controls how expected counts are calculated. Either: 1) A numeric value - used as a constant expected counts value for all query graphs (DEFAULT: 0). 2) A simplified igraph object - used as a reference graph from which expected counts are calculated for all query graphs. 3) NULL - Expected counts will be calculated based on the properties of the query graphs themselves.

max_graphlet_size

Generate graphlets up to this size. Currently only 4 and 5 are supported.

neighbourhood_size

Ego network neighbourhood size.

min_ego_nodes

Filter ego networks which have fewer than min_ego_nodes nodes.

min_ego_edges

Filter ego networks which have fewer than min_ego_edges edges.

binning_fn

Function used to bin ego network densities. Takes edge densities as its single argument, and returns a named list including, the input densities, the resulting bin breaks (vector of density bin limits), and the vector interval_indexes which states to what bin each of the individual elements in densities belongs to. ego network). If NULL, then the method binned_densities_adaptive with min_counts_per_interval = 5 and num_intervals = 100 is used (Default: NULL).

bin_counts_fn

Function used to calculate expected graphlet counts in each density bin. Takes graphlet_counts, interval_indexes (bin indexes) and max_graphlet_size as arguments. If bin_counts_fn is NULL, (default), it will apply either the approach from the original Netdis paper, or the respective Geometric-Poisson approximation; depending on the values of ref_graph and graphlet_counts_ref.

exp_counts_fn

Function used to map from binned reference counts to expected counts for each graphlet in each ego network of the query graphs. Takes ego_networks, density_bin_breaks, binned_graphlet_counts, and max_graphlet_size as arguments. If exp_counts_fn is NULL, (default), it will apply either the approach from the original Netdis paper, or the respective Geometric-Poisson approximation; depending on the values of ref_graph and graphlet_counts_ref.

graphlet_counts_1

Pre-generated graphlet counts for the first query graph. If the graphlet_counts_1 argument is defined then graph_1 will not be used.

graphlet_counts_compare

Named list of pre-generated graphlet counts for the remaining query graphs. If the graphlet_counts_compare argument is defined then graphs_compare will not be used.

graphlet_counts_ref

Pre-generated reference graphlet counts. If the graphlet_counts_ref argument is defined then ref_graph will not be used.

Value

Netdis statistics between graph_1 and graph_2 for graphlet sizes up to and including max_graphlet_size