Read all graphs in a directory, simplifying as requested

Reads graph data from all files in a directory matching the specified filename pattern. From each file, an a igraph graph object is constructed and the requested subset of the following simplifications is made in the following order: 1. Makes the graph undirected 2. Removes loops (where both endpoints of an edge are the same vertex) 3. Removes multiple edges (i.e. ensuring only one edge exists for each pair of endpoints) 4. Removes isolated vertices (i.e. vertices with no edges after the previous alterations)

read_simple_graphs(
  source_dir,
  format = "ncol",
  pattern = "*",
  as_undirected = TRUE,
  remove_loops = TRUE,
  remove_multiple = TRUE,
  remove_isolates = TRUE
)

Arguments

source_dir

Path to directory containing files with graph data

format

Format of graph data. Any format supported by igraph::read_graph can be used.

pattern

Pattern to use to filter filenames. Any pattern supported by dir can be used.

as_undirected

If TRUE make graph edges undirected

remove_loops

If TRUE, remove edgeds that connect a vertex to itself

remove_multiple

If TRUE remove multiple edges connencting the same pair of vertices

remove_isolates

If TRUE, remove vertices with no edges after the previous alterations have been made

Value

A named list of simplified igraph graph object, with the name of each graph set to the name of the file it was read from.

Examples

# Set source directory for Virus protein-protein interaction edge files stored in the netdist package.
source_dir <- system.file(file.path("extdata", "VRPINS"), package = "netdist")
print(source_dir) 
#> [1] "/home/runner/work/_temp/Library/netdist/extdata/VRPINS"
# Load query graphs as igraph objects
graph_1 <- read_simple_graph(file.path(source_dir, "EBV.txt"), format = "ncol")
graph_1
#> IGRAPH fa2b155 UN-- 60 208 -- 
#> + attr: name (v/c)
#> + edges from fa2b155 (vertex names):
#>  [1] A73  --BALF3 A73  --BARF0 A73  --BBLF2 A73  --BDRF1 A73  --BFRF4
#>  [6] A73  --BGLF2 A73  --BGLF3 A73  --BGLF5 A73  --BLLF2 A73  --BTRF1
#> [11] BALF3--BBLF2 BALF3--BDRF1 BALF3--BFRF4 BALF3--BGLF5 BALF3--BTRF1
#> [16] BALF3--BALF1 BALF3--BALF2 BALF3--BORF1 BALF3--BALF4 BALF3--BFLF2
#> [21] BALF3--BPLF1 BALF3--BALF5 BALF3--BBLF4 BALF3--BDLF2 BALF3--BdRF1
#> [26] BALF3--BERF3 BALF3--BHRF1 BALF3--LMP2A BARF0--BBLF2 BARF0--BFRF4
#> [31] BARF0--BSRF1 BARF0--BALF4 BARF0--BPLF1 BARF0--BALF5 BARF0--BDLF2
#> [36] BARF0--BdRF1 BARF0--BERF3 BARF0--BGLF1 BARF0--LMP2A BBLF2--BDRF1
#> + ... omitted several edges