Protein-protein interaction (PPI) networks for 5 microorganisms

A dataset containing the protein-protein interaction networks for the following 5 microorganisms

EBV

Common name: Epstein Barr virus

Scientific name: Human gammaherpesvirus 4

TaxonomyID: 10376

ECL

Common name: E.coli

Scientific name: Escherichia coli

TaxonomyID: 562

HSV-1

Common name: Herpes simplex virus type 1

Scientific name: Human alphaherpesvirus 1

TaxonomyID: 10298

KSHV

Common name: Karposi's Sarcoma-Associated Herpesvirus

Scientific name: Human gammaherpesvirus 8

TaxonomyID: 37296

VZV

Common name: Varicella zonster virus

Scientific name: Human alphaherpesvirus 3

TaxonomyID: 10335

virusppi

Format

A list of igraph objects.

Source

PPI data (EBV, HSV-1, KSHV, VZV): Fossum E, Friedel CC, Rajagopala SV, Titz B, Baiker A, Schmidt T, et al. (2009) Evolutionarily Conserved Herpesviral Protein Interaction Networks. PLoS Pathog 5(9): e1000570. https://doi.org/10.1371/journal.ppat.1000570. Data from Table S2 in the supporting information.

PPI data (ECL): Peregrín-Alvarez JM, Xiong X, Su C, Parkinson J (2009) The Modular Organization of Protein Interactions in Escherichia coli. PLoS Comput Biol 5(10): e1000523. https://doi.org/10.1371/journal.pcbi.1000523

Taxonomy ground truth: NCBI taxonomy database. https://www.ncbi.nlm.nih.gov/taxonomy