World trade networks from 1985–2014

The world trade data set consists of a small sample of world trade networks for the years 2001-2014, and pre-computed subgraph counts of a larger set of world trade networks (1985–2014). The world trade networks are based on the data set from [Feenstra et al., 2005] for the years 1962- 2000 and on the United Nations division COMTRADE [Division, 2015] for the years 2001-2014.

worldtradesub

Format

A list of two elements. The first element, 'wtnets', is a list of igraph objects providing a small sample of world trade networks from 2001–2014. The second element, 'Counts', is a list of pre-computed subgraph counts of world trade networks in the years 1985-2014.

Source

World trade networks:. United nations commodity trade statistics database (UN comtrade). http://comtrade.un.org/, 2015.

Subgraph Counts: Feenstra RC,Lipsey RE, Deng H, Ma AC, and Mo H. (2005) World trade flows: 1962-2000. Technical report, National Bureau of Economic Research. (See also https://cid.econ.ucdavis.edu/wix.html).

Details

wtnets: List of igraph objects providing the world trade networks from 2001–2014.

Counts: Pre-computed graphlet counts for the world trade networks in the years 1985-2014.